requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
gspread>=5.11.0

# Data processing
pandas>=2.1.0

# HTTP requests
requests>=2.31.0

# Web scraping
beautifulsoup4>=4.12.0

# AI/LLM
openai>=1.0.0
anthropic>=0.18.0

# Apify client (for lead scraping)
apify-client>=1.6.0

# Website scraping for GMaps pipeline
httpx>=0.25.0
html2text>=2020.1.16

# Add other dependencies as needed
